Hotfix: post-package-install event on all installed packages
As we noted yday during skeleton installation we are not prompted
for all packages to be injected into modules.

We were not prompted for in-direct dependencies, and I have discovered
in code that we submit manually event just for optional packages
and not dependencies.

As we are using different EventDispatcher (to avoid curcular calls) we need
broadcast post-package-install event to the original EventDispatcher
of the main composer process. This simplify the code a lot as we don't
need to create custom component installer.