fix: pass a string, not an array of strings, when providing a question prompt
Composer v1 allows a string or an array of strings as the first argument to `IOInterface::ask()`.
However, Composer v2 does not, and delegates directly to a symfony/console `Question` instance, which requires specifically a string.
Since we always passed an array with a single string, it was safe to change these values to strings throughout.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>